build: reorg files
