+++ src/app/components/StatusPicture/index.tsx
