+++ src/app/components/StatusPicture/styles.module.scss
