+++ src/app/views/ServerSettingsView/index.tsx
@@ -10,6 +10,22 @@ import {
   Text,
 } from "pavelLaptev/react-figma-ui/ui";
 
+type ViewsConfigI = {
+  [K in ServerType]: {
+    title: string;
+    description: React.ReactNode;
+    isEnabled: boolean;
+    fields: {
+      readonly id: string;
+      readonly type: "input" | "textarea" | "select";
+      readonly required: boolean;
+      readonly placeholder?: string;
+      readonly options?: string[];
+      value: string;
+    }[];
+  };
+};
+
 interface ViewProps {
   JSONsettingsConfig: JSONSettingsConfigI;
   setJSONsettingsConfig: React.Dispatch<
