+++ tsconfig.json
@@ -10,16 +10,8 @@
     "noEmit": true,
     "allowSyntheticDefaultImports": true,
     "lib": [
-      "es2021",
-      "dom"
+      "es2021"
     ],
-    "jsx": "react",
-    // "typeRoots": [
-    //   "./node_modules/@types",
-    //   "./node_modules/@figma",
-    //   "./node_modules/@tokens-bruecke",
-    //   "./node_modules/@figma/rest-api-spec"
-    // ]
   },
   "include": [],
   "files": [],
